Extract shared helpers from import_config_routes
First cut into blueprints/import_config_routes.py.

## Summary

blueprints/import_config_routes.py: 1384 -> 1188 (-196 / -14.2%)

## What moved

**New module**: blueprints/import_config_helpers.py (~244 lines)

Contains the 12 pure, side-effect-free helpers the four
/import-config/* routes share:

- _import_preview_json_default -- JSON serializer default
- _coerce_validation_response_payload -- unwrap validation resp
- _parse_csv_or_list_to_set -- CSV/list -> stripped set
- _parse_base_plex_libraries -- read library names from saved config
- 6 credential parsers (Plex + TMDb, from config/base/form)
- count_annotated_lines -- count YAML lines tagged imported/not_imported
- _map_playlist_libraries -- rewrite playlist library names through mapping

## Backward compatibility

blueprints/import_config_routes.py adds a top-level import from
the new helpers module and re-exports all 12 names.  This means:

- All existing 'from blueprints.import_config_routes import _xxx'
  still work
- quickstart.py's re-export chain (used by tests via qs_module.xxx)
  keeps working with no change
- Test suite: 0 failures

## Verification

- AST-verified: all 12 helper bodies byte-identical to develop
- AST-verified: all 4 route functions byte-identical to develop
- Backward-compat import test: quickstart.py re-exports point to
  the same objects as the new module
- Full test suite passes
- Ruff + black clean
- Dropped unused 're' import from routes file

## What's left in the routes file (1188 lines)

Still above the 600-line soft-limit:

- import_config_preview (~510 lines): the elephant.  Zip
  extraction, plex validation, tmdb validation are the three
  natural sub-blocks the module docstring flags for future
  decomposition.
- import_config_confirm (~354 lines)
- import_config_preview_mapped (~194 lines)
- import_config_report (~62 lines)

The docstring's file-size note has been updated to reflect the
extraction and note that the elephant is still the elephant.

## Sprint 5 scoreboard update

Coming into this PR: logscan_recommendations_engine.py had already
graduated (1429 -> 205 across 3 PRs).  import_config_routes.py was
the biggest remaining offender at 1384; now 1188.
